Move MPCNN API to common module (#105)
