+++ common/evaluation.py
@@ -1,11 +1,11 @@
-from mp_cnn.evaluators.sick_evaluator import SICKEvaluator
-from mp_cnn.evaluators.msrvid_evaluator import MSRVIDEvaluator
-from mp_cnn.evaluators.trecqa_evaluator import TRECQAEvaluator
-from mp_cnn.evaluators.wikiqa_evaluator import WikiQAEvaluator
+from .evaluators.sick_evaluator import SICKEvaluator
+from .evaluators.msrvid_evaluator import MSRVIDEvaluator
+from .evaluators.trecqa_evaluator import TRECQAEvaluator
+from .evaluators.wikiqa_evaluator import WikiQAEvaluator
 from nce.nce_pairwise_mp.evaluators.trecqa_evaluator import TRECQAEvaluatorNCE
 from nce.nce_pairwise_mp.evaluators.wikiqa_evaluator import WikiQAEvaluatorNCE
 
-class MPCNNEvaluatorFactory(object):
+class EvaluatorFactory(object):
     """
     Get the corresponding Evaluator class for a particular dataset.
     """
@@ -27,9 +27,9 @@ class MPCNNEvaluatorFactory(object):
             return None
 
         if nce:
-            evaluator_map = MPCNNEvaluatorFactory.evaluator_map_nce
+            evaluator_map = EvaluatorFactory.evaluator_map_nce
         else:
-            evaluator_map = MPCNNEvaluatorFactory.evaluator_map
+            evaluator_map = EvaluatorFactory.evaluator_map
 
         if not hasattr(dataset_cls, 'NAME'):
             raise ValueError('Invalid dataset. Dataset should have NAME attribute.')
