+++ common/evaluators/sick_evaluator.py
@@ -1,19 +1,19 @@
-from scipy.stats import pearsonr
+from scipy.stats import pearsonr, spearmanr
 import torch
 import torch.nn.functional as F
 
-from mp_cnn.evaluators.evaluator import Evaluator
+from .evaluator import Evaluator
 
 
-class MSRVIDEvaluator(Evaluator):
+class SICKEvaluator(Evaluator):
 
     def __init__(self, dataset_cls, model, data_loader, batch_size, device):
-        super(MSRVIDEvaluator, self).__init__(dataset_cls, model, data_loader, batch_size, device)
+        super(SICKEvaluator, self).__init__(dataset_cls, model, data_loader, batch_size, device)
 
     def get_scores(self):
         self.model.eval()
         num_classes = self.dataset_cls.NUM_CLASSES
-        predict_classes = torch.arange(0, num_classes).expand(self.batch_size, num_classes)
+        predict_classes = torch.arange(1, num_classes + 1).expand(self.batch_size, num_classes)
         test_kl_div_loss = 0
         predictions = []
         true_labels = []
@@ -23,7 +23,7 @@ class MSRVIDEvaluator(Evaluator):
             test_kl_div_loss += F.kl_div(output, batch.label, size_average=False).data[0]
             # handle last batch which might have smaller size
             if len(predict_classes) != len(batch.sentence_1):
-                predict_classes = torch.arange(0, num_classes).expand(len(batch.sentence_1), num_classes)
+                predict_classes = torch.arange(1, num_classes + 1).expand(len(batch.sentence_1), num_classes)
 
             if self.data_loader.device != -1:
                 with torch.cuda.device(self.device):
@@ -38,5 +38,6 @@ class MSRVIDEvaluator(Evaluator):
         true_labels = torch.cat(true_labels).cpu().numpy()
         test_kl_div_loss /= len(batch.dataset.examples)
         pearson_r = pearsonr(predictions, true_labels)[0]
+        spearman_r = spearmanr(predictions, true_labels)[0]
 
-        return [pearson_r, test_kl_div_loss], ['pearson_r', 'KL-divergence loss']
+        return [pearson_r, spearman_r, test_kl_div_loss], ['pearson_r', 'spearman_r', 'KL-divergence loss']
