+++ common/trainers/qa_trainer.py
@@ -4,7 +4,7 @@ import torch
 import torch.nn.functional as F
 from torch.optim.lr_scheduler import ReduceLROnPlateau
 
-from mp_cnn.trainers.trainer import Trainer
+from .trainer import Trainer
 
 
 class QATrainer(Trainer):
